database/pgsql: move token lib
